Add files via upload
Snake, Water and Gun or Rock, Paper and Scissor is a game most of you have played in school time.(I some time play even now) This is a program capable playing the game with you. This program able to print the result after you choose snake, water or gun.